backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
transformers>=4.40.0
torch>=2.2.0
torchvision>=0.17.0
Pillow>=10.0.0
opencv-python-headless>=4.9.0
numpy>=1.26.0